requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
plotly==5.18.0
openpyxl==3.1.2
prophet==1.1.5
statsmodels==0.14.1
scikit-learn==1.3.2
tensorflow==2.15.0
matplotlib==3.8.2
seaborn==0.13.0